RoadSage/rpi/requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
ultralytics==8.2.0
tflite-runtime==2.14.0
adafruit-circuitpython-neopixel==6.3.11
bleak==0.22.3
pyserial==3.5
